+++ packages/vite-plugin/vitest.config.ts
@@ -6,5 +6,7 @@ export default defineConfig({
       // See https://github.com/webdiscus/ansis/?tab=readme-ov-file#disable-colors-in-tests
       NO_COLOR: 'true',
     },
+    // Some of this package's tests are slow on Windows
+    testTimeout: 30_000,
   },
 })