+++ packages/edge-functions/dev/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [1.0.10](https://github.com/netlify/primitives/compare/edge-functions-dev-v1.0.9...edge-functions-dev-v1.0.10) (2026-02-06)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/runtime-utils bumped from 2.2.1 to 2.3.0
+
+## [1.0.9](https://github.com/netlify/primitives/compare/edge-functions-dev-v1.0.8...edge-functions-dev-v1.0.9) (2026-01-28)
+
+
+### Bug Fixes
+
+* **deps:** update netlify packages ([#576](https://github.com/netlify/primitives/issues/576)) ([1c6ff24](https://github.com/netlify/primitives/commit/1c6ff247318b8cdb1961ce76e96fe423a4ec8cd7))
+
+## [1.0.8](https://github.com/netlify/primitives/compare/edge-functions-dev-v1.0.7...edge-functions-dev-v1.0.8) (2026-01-19)
+
+
+### Bug Fixes
+
+* **deps:** update netlify packages ([#566](https://github.com/netlify/primitives/issues/566)) ([184656a](https://github.com/netlify/primitives/commit/184656af871a81bddc69e9133f22bfa945f28ebf))
+
 ## [1.0.7](https://github.com/netlify/primitives/compare/edge-functions-dev-v1.0.6...edge-functions-dev-v1.0.7) (2025-12-08)