+++ packages/nuxt-module/CHANGELOG.md
@@ -1,5 +1,41 @@
 # Changelog
 
+## [0.2.21](https://github.com/netlify/primitives/compare/nuxt-v0.2.20...nuxt-v0.2.21) (2026-02-06)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from ^4.9.0 to ^4.10.0
+
+## [0.2.20](https://github.com/netlify/primitives/compare/nuxt-v0.2.19...nuxt-v0.2.20) (2026-01-28)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from ^4.8.8 to ^4.9.0
+
+## [0.2.19](https://github.com/netlify/primitives/compare/nuxt-v0.2.18...nuxt-v0.2.19) (2026-01-19)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from ^4.8.7 to ^4.8.8
+
+## [0.2.18](https://github.com/netlify/primitives/compare/nuxt-v0.2.17...nuxt-v0.2.18) (2025-12-21)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from ^4.8.6 to ^4.8.7
+
 ## [0.2.17](https://github.com/netlify/primitives/compare/nuxt-v0.2.16...nuxt-v0.2.17) (2025-12-17)